requirements.txt
requests
google-genai
python-dotenv
dataclasses-json
typing-extensions
Pillow
